analysis/requirements.txt
azure-storage-blob
pandas
nltk
transformers
torch
dotenv
azure-data-tables